fixed problems with last PR
