+++ .gitignore
@@ -1,6 +1,8 @@
 # IDE
 .idea
 
+*.pdf
+
 # binaries
 .database
 .bin
